% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{check_ecotox_availability}
\alias{check_ecotox_availability}
\title{Check whether a ECOTOX database exists locally}
\usage{
check_ecotox_availability(target = get_ecotox_path())
}
\arguments{
\item{target}{A \code{character} string specifying the path where to look for the database file.}
}
\value{
Returns a \code{logical} value indicating whether a copy of the database exists. It also returns
a \code{files} attribute that lists which copies of the database are found.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} Tests whether a local copy of the US EPA ECOTOX database exists in
\code{\link[=get_ecotox_path]{get_ecotox_path()}}.
}
\details{
When arguments are omitted, this function will look in the default directory (\code{\link[=get_ecotox_path]{get_ecotox_path()}}).
However, it is possible to build a database file elsewhere if necessary.
}
\examples{
check_ecotox_availability()
}
\seealso{
Other database-access-functions: 
\code{\link{check_ecotox_build}()},
\code{\link{check_ecotox_version}()},
\code{\link{cite_ecotox}()},
\code{\link{dbConnectEcotox}()},
\code{\link{get_ecotox_info}()},
\code{\link{get_ecotox_sqlite_file}()},
\code{\link{list_ecotox_fields}()}
}
\author{
Pepijn de Vries
}
\concept{database-access-functions}
