% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{migrate_ecotox_path}
\alias{migrate_ecotox_path}
\title{Temporary function to switch to new ECOTOX path}
\usage{
migrate_ecotox_path(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical} value. Should files that already exist in the
new location be overwritten?}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Since ECOTOXr v1.2.5, the path is located with the base package \code{tools}
instead of the previously used \code{rappdirs}. Use this function to migrate
your old files to this new location.
}
\seealso{
Other database-access-functions:
\code{\link[=check_ecotox_availability]{check_ecotox_availability()}},
\code{\link[=check_ecotox_build]{check_ecotox_build()}},
\code{\link[=check_ecotox_version]{check_ecotox_version()}},
\code{\link[=cite_ecotox]{cite_ecotox()}},
\code{\link[=dbConnectEcotox]{dbConnectEcotox()}},
\code{\link[=get_ecotox_info]{get_ecotox_info()}},
\code{\link[=get_ecotox_sqlite_file]{get_ecotox_sqlite_file()}},
\code{\link[=list_ecotox_fields]{list_ecotox_fields()}}
}
\concept{database-access-functions}
