% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.r
\name{as_unit_ecotox}
\alias{as_unit_ecotox}
\title{Values represented by ECOTOX \code{character} to units}
\usage{
as_unit_ecotox(
  x,
  type = c("concentration", "duration", "length", "media", "application", "size",
    "weight"),
  ...,
  warn = TRUE
)
}
\arguments{
\item{x}{A vector of \code{character} strings. It expects fields as commonly returned
from the ECOTOX database.}

\item{type}{The type of unit that can help the sanitation process. See the 'usage'
section for available options. These options are linked to the different unit tables
in the database (see \code{vignette("ecotox-schema")}). It can help to interpret ambiguous
units correctly. For instance, 'dpm' can both mean 'disintegrations per minute'
(\code{type = "concentration"}) and 'days post-moult' (\code{type = "duration"}).}

\item{...}{TODO}

\item{warn}{If set to \code{FALSE} warnings while converting text to units are suppressed.}
}
\value{
A vector of \code{?units::unit} class objects with the same length as \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Convert text to units after
sanitising.
}
\details{
The following steps are performed (in the order as listed)
to sanitise text before coercing it to units:
\itemize{
\item TODO
}

It is your own responsibility to check if the sanitising steps are appropriate for
your analyses.
}
\examples{
## TODO
}
\seealso{
Other ecotox-sanitisers: 
\code{\link{as_date_ecotox}()},
\code{\link{as_numeric_ecotox}()},
\code{\link{process_ecotox_dates}()},
\code{\link{process_ecotox_numerics}()},
\code{\link{process_ecotox_units}()}
}
\author{
Pepijn de Vries
}
\concept{ecotox-sanitisers}
