% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{migrate_ecotox_path}
\alias{migrate_ecotox_path}
\title{Temporary function to switch to new ECOTOX path}
\usage{
migrate_ecotox_path(overwrite = FALSE)
}
\arguments{
\item{overwrite}{\code{logical} value. Should files that already exist in the
new location be overwritten?}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Since ECOTOXr v1.2.5, the path is located with the base package \code{tools}
instead of the previously used \code{rappdirs}. Use this function to migrate
your old files to this new location.
}
